+++ resolwe_bio/tests/processes/test_pca.py
@@ -53,17 +53,62 @@ class PcaProcessorTestCase(KBBioProcessTestCase):
                 source="DICTYBASE",
                 species="Dictyostelium discoideum",
             )
+            expression_5 = self.prepare_expression(
+                f_rc="exp_6_rc.tab.gz",
+                f_exp="exp_6_tpm.tab.gz",
+                f_type="TPM",
+                source="DICTYBASE",
+                species="Dictyostelium discoideum",
+            )
+            expression_6 = self.prepare_expression(
+                f_rc="exp_7_rc.tab.gz",
+                f_exp="exp_7_tpm.tab.gz",
+                f_type="TPM",
+                source="DICTYBASE",
+                species="Dictyostelium discoideum",
+            )
+            expression_7 = self.prepare_expression(
+                f_rc="exp_8_rc.tab.gz",
+                f_exp="exp_8_tpm.tab.gz",
+                f_type="TPM",
+                source="DICTYBASE",
+                species="Dictyostelium discoideum",
+            )
 
         inputs = {
-            "exps": [expression_1.pk, expression_2.pk],
+            "exps": [
+                expression_1.pk,
+                expression_2.pk,
+                expression_5.pk,
+                expression_6.pk,
+                expression_7.pk,
+            ],
             "source": "DICTYBASE",
             "species": "Dictyostelium discoideum",
         }
         pca = self.run_process("pca", inputs)
         saved_json, test_json = self.get_json("pca_plot.json.gz", pca.output["pca"])
+        # returns 4 PCA components. Last component differs when testing on different systems
+        # and is not tested here.
         self.assertAlmostEqualGeneric(
-            round_elements(test_json["flot"]["data"]),
-            round_elements(saved_json["flot"]["data"]),
+            round_elements(test_json["flot"]["data"][0][:3]),
+            round_elements(saved_json["flot"]["data"][0][:3]),
+        )
+        self.assertAlmostEqualGeneric(
+            round_elements(test_json["flot"]["data"][1][:3]),
+            round_elements(saved_json["flot"]["data"][1][:3]),
+        )
+        self.assertAlmostEqualGeneric(
+            round_elements(test_json["flot"]["data"][2][:3]),
+            round_elements(saved_json["flot"]["data"][2][:3]),
+        )
+        self.assertAlmostEqualGeneric(
+            round_elements(test_json["flot"]["data"][3][:3]),
+            round_elements(saved_json["flot"]["data"][3][:3]),
+        )
+        self.assertAlmostEqualGeneric(
+            round_elements(test_json["flot"]["data"][4][:3]),
+            round_elements(saved_json["flot"]["data"][4][:3]),
         )
         self.assertAlmostEqualGeneric(
             round_elements(test_json["explained_variance_ratios"]),
@@ -84,8 +129,8 @@ class PcaProcessorTestCase(KBBioProcessTestCase):
         pca = self.run_process("pca", inputs)
         saved_json, test_json = self.get_json("pca_plot_2.json.gz", pca.output["pca"])
         self.assertAlmostEqualGeneric(
-            round_elements(test_json["flot"]["data"]),
-            round_elements(saved_json["flot"]["data"]),
+            round_elements(test_json["flot"]["data"][0][:3]),
+            round_elements(saved_json["flot"]["data"][0][:3]),
         )
 
         self.assertEqual(len(pca.process_warning), 0)
@@ -95,8 +140,8 @@ class PcaProcessorTestCase(KBBioProcessTestCase):
         pca = self.run_process("pca", inputs)
         saved_json, test_json = self.get_json("pca_plot_3.json.gz", pca.output["pca"])
         self.assertAlmostEqualGeneric(
-            round_elements(test_json["flot"]["data"]),
-            round_elements(saved_json["flot"]["data"]),
+            round_elements(test_json["flot"]["data"][0][:3]),
+            round_elements(saved_json["flot"]["data"][0][:3]),
         )
 
         self.assertEqual(len(pca.process_warning), 0)